LayoutTests/imported/w3c/web-platform-tests/eventsource/format-utf-8.any-expected.txt

CONSOLE MESSAGE: EventSource's response has a charset ("windows-1252") that is not UTF-8. The response will be decoded as UTF-8.

PASS EventSource always UTF-8
